blog_manager/requirements.top.txt
autoflake==2.3.1
black==25.1.0
cryptography==45.0.6
django-anymail==13.0.1
django-cloudinary-storage==0.3.0
django-cors-headers==4.7.0
django-environ==0.12.0
django-filter==25.1
drf-spectacular==0.28.0
flake8==7.3.0
git-filter-repo==2.47.0
iniconfig==2.1.0
isort==6.0.1
mysqlclient==2.2.7
pillow==11.3.0
pip==25.2
pre_commit==4.3.0
psycopg==3.2.10
psycopg-binary==3.2.10
PyGithub==2.8.1
Pygments==2.19.2
pytest-cov==6.2.1
pytest-django==4.11.1
python-dotenv==1.1.1
